Renamed modules to core modules
