+++ Core/src/script/CoreModules.cpp
@@ -1,5 +1,5 @@
 
-#include "Modules.h"
+#include "CoreModules.h"
 
 #include "gui/Gui.h"
 #include "gui/GuiRegistry.h"
